Bugfix: If the DMA transfer isn't done, wait
The issue was: the DMA transfer was sometimes being manually stopped too quickly, before it had finished.
This resulted in various glitchy effects: a sort of screen tearing in 1-Screen Mode, and parts of screens appearing "glitched together" in 2-Screen Mode (appearing corrupted as a bonus when the top and bottom screen aren't the same color format).
Those glitches should be fixed now, for the most part.